# api/openapi.yaml
openapi: 3.1.0
info:
  title: betledger API
  version: 0.1.0
  description: |
    Wallet and wagering operations for game providers.

    Monetary amounts are always decimal strings with at most two decimal places,
    never JSON numbers, so that values never pass through floating point.

    Every response carries `X-Correlation-Id`. A value sent by the client is kept
    when it has 1 to 128 characters among letters, digits, `-`, `_` and `.`;
    otherwise a new one is generated.

    Business endpoints require an OAuth 2.0 access token issued by the identity
    provider through the `client_credentials` grant, sent as
    `Authorization: Bearer <token>`. The token must be signed by the provider,
    unexpired, and carry the `betledger-api` audience. Health endpoints are
    public.

    Authorization uses realm roles from the token: opening wallets requires
    `wallet-operator`, granted only to internal services, and sending operations
    requires `game-provider`. A provider acts only for the provider in its
    `provider_id` claim.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /wallets:
    post:
      tags: [Wallets]
      operationId: openWallet
      summary: Open a wallet
      description: |
        Opens a wallet for a player in a currency. A player holds at most one
        wallet per currency. Requires the `wallet-operator` role.

        With a positive initial balance, the wallet, the `OPENING` transaction,
        its credit ledger entry and the `WagerTransactionProcessed` and
        `WalletBalanceChanged` events are committed together. With a zero initial
        balance, only the wallet is created.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenWalletRequest'
            example:
              playerId: 0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1
              initialBalance:
                amount: '1000.00'
                currency: BRL
      responses:
        '201':
          description: Wallet opened.
          headers:
            Location:
              description: Path of the created wallet.
              schema:
                type: string
                example: /wallets/0192f291-27dd-7d3f-8071-5f8685deef37
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: The player already holds a wallet in this currency.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: WALLET_ALREADY_EXISTS
                  message: player 0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1 already holds a BRL wallet
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /wallets/{walletId}:
    get:
      tags: [Wallets]
      operationId: getWallet
      summary: Read a wallet
      description: Returns the current state of the wallet. Requires the `wallet-operator` role.
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The wallet.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/WalletNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /wallets/{walletId}/ledger:
    get:
      tags: [Wallets]
      operationId: getWalletLedger
      summary: Read the wallet ledger
      description: |
        Returns the ledger entries of the wallet, oldest first, in a stable
        order. Paging uses an opaque cursor: send back the `nextCursor` of a page
        to read the next one. The last page carries no `nextCursor`. Requires the
        `wallet-operator` role.
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - name: cursor
          in: query
          required: false
          description: The `nextCursor` of the previous page.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: How many entries to return.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: A page of the ledger.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerPage'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/WalletNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /wallets/{walletId}/reconciliation:
    post:
      tags: [Wallets]
      operationId: reconcileWallet
      summary: Reconcile a wallet against its ledger
      description: |
        Rebuilds the balance from every ledger entry of the wallet, including the
        opening, and compares it with the stored balance, both read in a single
        consistent view. `difference` is the stored balance minus the rebuilt
        one. Reconciling never changes a balance: a divergence is reported here
        and logged, not corrected. Requires the `wallet-operator` role.
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The comparison between the stored balance and the rebuilt one.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reconciliation'
              example:
                walletId: 0192f291-27dd-7d3f-8071-5f8685deef37
                storedBalance: { amount: '975.00', currency: BRL }
                calculatedBalance: { amount: '975.00', currency: BRL }
                difference: { amount: '0.00', currency: BRL }
                consistent: true
                checkedEntries: 2
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/WalletNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token obtained from the identity provider with the
        `client_credentials` grant. Locally:

        ```sh
        curl -s http://localhost:8081/realms/betledger/protocol/openid-connect/token \
          -d grant_type=client_credentials \
          -d client_id=provider-a -d client_secret=provider-a-secret
        ```
  parameters:
    WalletId:
      name: walletId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Identifier that ties together logs and events of the request.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
        example: req-123
...
  headers:
    CorrelationId:
      description: The correlation identifier kept or generated for the request.
      schema:
        type: string
  responses:
    InvalidInput:
      description: The request is invalid and must be fixed before being sent again.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalidAmount:
              summary: Amount with more than two decimal places
              value:
                error:
                  code: INVALID_AMOUNT
                  message: monetary amount "10.123" exceeds the scale of 2 decimal places
            invalidInput:
              summary: Malformed body
              value:
                error:
                  code: INVALID_INPUT
                  message: 'malformed request body: unexpected EOF'
    Unauthenticated:
      description: The bearer token is missing, invalid, expired or issued for another audience.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer realm="betledger"
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: a bearer token is required
    WalletNotFound:
      description: The wallet does not exist.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: WALLET_NOT_FOUND
              message: wallet 0192f291-27dd-7d3f-8071-5f8685deef37 not found
...
    Forbidden:
      description: |
        The caller is authenticated but lacks the role for the endpoint, or the
        operation names a provider other than the one in the token.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: the wallet-operator role is required
    ServiceUnavailable:
      description: A dependency is temporarily unavailable. The request can be retried.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: service temporarily unavailable, retry later
    InternalError:
      description: Unexpected failure. Details are logged, never returned.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: internal error
  schemas:
    MoneyInput:
      type: object
      description: Non-negative amount accepted in requests.
      required: [amount, currency]
      additionalProperties: false
      properties:
        amount:
          type: string
          description: Decimal string with at most two decimal places.
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
          example: '1000.00'
        currency:
          type: string
          description: ISO 4217 code.
          pattern: '^[A-Z]{3}$'
          example: BRL
    Money:
      type: object
      description: Amount returned by the service, always with two decimal places.
      required: [amount, currency]
      additionalProperties: false
      properties:
        amount:
          type: string
          pattern: '^-?[0-9]+\.[0-9]{2}$'
          example: '1000.00'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: BRL
    OpenWalletRequest:
      type: object
      required: [playerId, initialBalance]
      additionalProperties: false
      properties:
        playerId:
          type: string
          format: uuid
        initialBalance:
          $ref: '#/components/schemas/MoneyInput'
    Wallet:
      type: object
      required: [id, playerId, balance, version]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        playerId:
          type: string
          format: uuid
        balance:
          $ref: '#/components/schemas/Money'
        version:
          type: integer
          format: int64
          minimum: 1
          description: Starts at 1 and increments on each balance change.
...
    LedgerEntry:
      type: object
      required: [id, transactionId, direction, money, balanceBefore, balanceAfter, recordedAt]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        money:
          $ref: '#/components/schemas/Money'
        balanceBefore:
          $ref: '#/components/schemas/Money'
        balanceAfter:
          $ref: '#/components/schemas/Money'
        recordedAt:
          type: string
          format: date-time
    LedgerPage:
      type: object
      required: [entries]
      additionalProperties: false
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntry'
        nextCursor:
          type: string
          description: Opaque cursor of the next page; absent on the last one.
    Reconciliation:
      type: object
      required: [walletId, storedBalance, calculatedBalance, difference, consistent, checkedEntries]
      additionalProperties: false
      properties:
        walletId:
          type: string
          format: uuid
        storedBalance:
          $ref: '#/components/schemas/Money'
        calculatedBalance:
          $ref: '#/components/schemas/Money'
        difference:
          $ref: '#/components/schemas/Money'
          description: The stored balance minus the rebuilt one.
        consistent:
          type: boolean
        checkedEntries:
          type: integer
          minimum: 0
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                Stable failure code, part of the contract. Codes returned so far:
                `INVALID_INPUT`, `INVALID_AMOUNT`, `AMOUNT_OVERFLOW`,
                `TRANSACTION_KIND_NOT_ALLOWED`, `WALLET_ALREADY_EXISTS`,
                `WALLET_NOT_FOUND`, `IDEMPOTENCY_CONFLICT`, `UNAUTHENTICATED`, `FORBIDDEN`,
                `SERVICE_UNAVAILABLE`,
                `INTERNAL_ERROR`.
            message:
              type: string
              description: Human-readable description; not meant to be parsed.
...